+++ src/routes/user/index.ts
@@ -11,10 +11,10 @@ const router = Router();
 router.post('/', controller.saveUser);
 router.patch('/', helloWorld);
 router.delete('/', helloWorld);
+
 router.use('/:id/role', roleRouter);
 router.use('/notification', notificationRouter);
 router.use('/category', categoryRouter);
-
 router.use('/profile', profileRouter);
 router.use('/login', loginRouter);
